Update README.md
This write up is a guide on how to schedule a task/ program to run at a given time.  As described in this writeup, a user might decide to keep note of their activities every 3 hours and a Task Scheduler is needed to achieve this. Also, a user might want to run an antivirus software at a given time .